Iraq and Afghanistan SIGACTS FOIA'd by Vincent Bauer

sigacts_afghanistan_df

"This data source publishes over 600,000 reports of ‘significant activities’ in Afghanistan from January 2008 through December 2014 These data were declassified by CENTCOM in 2014 and include information on both insurgent and Coalition activity. While freely available, these data are not usable by in their original form. I walk through the process of converting the declassified PDF documents into a dataframe and make the final results available as an Excel file. I have converted the military grid reference system (MGRS) locations to latitude and longtitude."

sigacts_iraq_df

"This data source publishes over 250,000 declassified reports of ‘significant activities’ in Iraq from January 2004 to July 2007. While freely available, these data are not usable by in their original form. I walk through the process of converting the declassified PDF documents into a dataframe and make the final results available as an Excel file. Unlike the data available for Afghanistan, these data do not include significant activities carried out by Coalition forces and are not geolocated."

sigacts_afghanistan_df

sigacts_iraq_df

Format

An object of class tbl_df (inherits from tbl, data.frame) with 431547 rows and 6 columns.

Source

https://stanford.edu/~vbauer/data.html

Examples

# Afghanistan =========================================================================
sigacts_afghanistan_df

#> # A tibble: 431,547 x 6
#>    time                event_type    event_category    mgrs    lat      lon     
#>    <dttm>              <chr>         <chr>             <chr>   <chr>    <chr>   
#>  1 2008-01-01 00:00:00 Enemy Action  Indirect Fire     41SNV7… 35.8352… 063.841…
#>  2 2008-01-01 00:45:00 Enemy Action  Direct Fire       41RQR7… 31.9001… 065.886…
#>  3 2008-01-01 01:20:00 Enemy Action  Indirect Fire     42SVD9… 34.8372… 068.945…
#>  4 2008-01-01 02:05:00 Explosive Ha… IED Found and Cl… 42SUA1… 31.6194… 66.9968…
#>  5 2008-01-01 04:00:00 Other         Planned Event     42SVD8… 35.2340… 068.879…
#>  6 2008-01-01 04:30:00 Explosive Ha… UNEXPLODED ORDNA… 42SVB8… 32.7356… 068.807…
#>  7 2008-01-01 05:08:00 Enemy Action  Direct Fire       42SXD7… 35.0201… 070.940…
#>  8 2008-01-01 05:13:00 Explosive Ha… IED Found and Cl… 41SQS7… 32.7022… 065.901…
#>  9 2008-01-01 05:30:00 Explosive Ha… UNEXPLODED ORDNA… 42SUB7… 32.5297… 067.615…
#> 10 2008-01-01 06:14:00 Explosive Ha… IED False         42SWB4… 33.0325… 069.481…
#> # … with 431,537 more rows

dplyr::glimpse(sigacts_afghanistan_df)

#> Observations: 431,547
#> Variables: 6
#> $ time           <dttm> 2008-01-01 00:00:00, 2008-01-01 00:45:00, 2008-01-01 …
#> $ event_type     <chr> "Enemy Action", "Enemy Action", "Enemy Action", "Explo…
#> $ event_category <chr> "Indirect Fire", "Direct Fire", "Indirect Fire", "IED …
#> $ mgrs           <chr> "41SNV7666", "41RQR7333", "42SVD9555", "42SUA1000", "4…
#> $ lat            <chr> "35.83523N", "31.90011N", "34.83728N", "31.6194718964N…
#> $ lon            <chr> "063.84149E", "065.88659E", "068.94531E", "66.99688204…

# Iraq ================================================================================ 
sigacts_iraq_df

#> # A tibble: 162,852 x 4
#>    time                event_type target                city      
#>    <dttm>              <chr>      <chr>                 <chr>     
#>  1 2004-01-01 00:20:00 IED        Coalition Forces      Baghdad   
#>  2 2004-01-01 00:45:00 Small Arms Iraqi Security Forces NA        
#>  3 2004-01-01 01:01:00 RPG        Coalition Forces      Ramadi    
#>  4 2004-01-01 01:11:00 Mortar     Coalition Forces      Abu Ghraib
#>  5 2004-01-01 01:15:00 IED        Coalition Forces      Abu Ghraib
#>  6 2004-01-01 01:15:00 Mortar     Coalition Forces      NA        
#>  7 2004-01-01 01:22:00 VBIED      Civilian              Baghdad   
#>  8 2004-01-01 02:00:00 Other      NA                    Mosul     
#>  9 2004-01-01 02:06:00 Mortar     Coalition Forces      Kirkuk    
#> 10 2004-01-01 02:25:00 Rocket     Coalition Forces      Baghdad   
#> # … with 162,842 more rows

dplyr::glimpse(sigacts_iraq_df)

#> Observations: 162,852
#> Variables: 4
#> $ time       <dttm> 2004-01-01 00:20:00, 2004-01-01 00:45:00, 2004-01-01 01:0…
#> $ event_type <chr> "IED", "Small Arms", "RPG", "Mortar", "IED", "Mortar", "VB…
#> $ target     <chr> "Coalition Forces", "Iraqi Security Forces", "Coalition Fo…
#> $ city       <chr> "Baghdad", NA, "Ramadi", "Abu Ghraib", "Abu Ghraib", NA, "…