The CNS Missile and SLV Launch Databases

dprk_missile_tests_df

The James Martin Center for Nonproliferation Studies (CNS) North Korea Missile Test Database is the first database to record flight tests of all missiles launched by North Korea capable of delivering a payload of at least 500 kilograms (1102.31 pounds) a distance of at least 300 kilometers (186.4 miles). The database captures advancements in North Korea's missile program by documenting all such tests since the first one occurred in April 1984, and will be routinely updated as events warrant.

iran_missile_tests_df

The James Martin Center for Nonproliferation Studies Iran Missile and SLV Launch Database records flight tests and operational uses of all missiles and SLVs launched by Iran that are capable of delivering a payload of at least 500 kilograms (1102.31 pounds) a distance of at least 300 kilometers (186.4 miles). The database captures advancements in Iran’s rocket programs by documenting all such launches by Iran since the end of the Iran-Iraq War in 1988. The database will be routinely updated as events warrant.

india_missile_tests_df and pakistan_missile_tests_df

The James Martin Center for Nonproliferation Studies (CNS) India and Pakistan Missile Launch Databases record flight tests of all missiles launched by India and Pakistan, as well as space launch vehicles (SLVs) tested by India, that are capable of delivering a payload of at least 500 kilograms (1102.31 pounds) or traveling a distance of at least 300 kilometers (186.4 miles). The database captures advancements in both countries’ rocket programs by documenting all such launches carried out by India and Pakistan. The database will be routinely updated as events warrant.

dprk_missile_tests_df

iran_missile_tests_df

india_missile_tests_df

pakistan_missile_tests_df

Format

An object of class tbl_df (inherits from tbl, data.frame) with 140 rows and 16 columns.

Source

https://www.nti.org/analysis/reports/cns-missile-and-slv-launch-databases/

References

Created by The James Martin Center for Nonproliferation Studies for the Nuclear Threat Initiative

Examples

# dprk ================================================================================
dprk_missile_tests_df

#> # A tibble: 140 x 16
#>    launch_time         missile_name missile_type launch_authority facility_name
#>    <dttm>              <chr>        <chr>        <chr>            <chr>        
#>  1 1984-04-08 16:00:00 Scud-B       SRBM         NA               Tonghae Sate…
#>  2 1984-04-08 16:00:00 Scud-B       SRBM         NA               Tonghae Sate…
#>  3 1984-04-08 16:00:00 Scud-B       SRBM         NA               Tonghae Sate…
#>  4 1984-08-31 17:00:00 Scud-B       SRBM         NA               Tonghae Sate…
#>  5 1984-08-31 17:00:00 Scud-B       SRBM         NA               Tonghae Sate…
#>  6 1984-08-31 17:00:00 Scud-B       SRBM         NA               Tonghae Sate…
#>  7 1986-04-30 17:00:00 NA           NA           NA               Tonghae Sate…
#>  8 1990-04-30 17:00:00 Nodong       MRBM         NA               Tonghae Sate…
#>  9 1990-05-31 17:00:00 Scud-C       SRBM         NA               Tonghae Sate…
#> 10 1991-06-30 17:00:00 Scud-C       SRBM         NA               Chihari Miss…
#> # … with 130 more rows, and 11 more variables: facility_location <chr>,
#> #   other_name <chr>, facility_latitude <chr>, facility_longitude <chr>,
#> #   landing_location <chr>, apogee_km <dbl>, distance_travelled_km <dbl>,
#> #   test_confirmed <lgl>, test_successful <lgl>, additional_information <chr>,
#> #   sources <chr>

dplyr::glimpse(dprk_missile_tests_df)

#> Observations: 140
#> Variables: 16
#> $ launch_time            <dttm> 1984-04-08 16:00:00, 1984-04-08 16:00:00, 198…
#> $ missile_name           <chr> "Scud-B", "Scud-B", "Scud-B", "Scud-B", "Scud-…
#> $ missile_type           <chr> "SRBM", "SRBM", "SRBM", "SRBM", "SRBM", "SRBM"…
#> $ launch_authority       <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ facility_name          <chr> "Tonghae Satellite Launching Ground", "Tonghae…
#> $ facility_location      <chr> "Hwadae County, North Hamgyong Province", "Hwa…
#> $ other_name             <chr> "Musudan-ri", "Musudan-ri", "Musudan-ri", "Mus…
#> $ facility_latitude      <chr> "40.849996599999997", "40.849996599999997", "4…
#> $ facility_longitude     <chr> "129.666664", "129.666664", "129.666664", "129…
#> $ landing_location       <chr> NA, NA, NA, NA, NA, NA, NA, NA, "Sea of Japan …
#> $ apogee_km              <dbl> NA, NA, 200, NA, 200, 200, NA, 150, 200, NA, 0…
#> $ distance_travelled_km  <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 50…
#> $ test_confirmed         <lgl> TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, FALSE, FAL…
#> $ test_successful        <lgl> TRUE, TRUE, TRUE, FALSE, FALSE, FALSE, NA, FAL…
#> $ additional_information <chr> "First known test by North Korea of a missile …
#> $ sources                <chr> "http://www.nti.org/media/pdfs/north_korea_mis…

# iran ================================================================================ 
iran_missile_tests_df

#> # A tibble: 138 x 14
#>    launch_time         missile_name facility_name facility_latitu…
#>    <dttm>              <chr>        <chr>         <chr>           
#>  1 1991-04-30 17:00:00 Shahab-2     NA            NA              
#>  2 1994-11-05 16:00:00 Shahab-1     Kermanshah A… 34.3458         
#>  3 1994-11-05 16:00:00 Shahab-1     Kermanshah A… 34.3458         
#>  4 1994-11-05 16:00:00 Shahab-1     Kermanshah A… 34.3458         
#>  5 1994-11-08 16:00:00 Shahab-1     Kermanshah A… 34.3458         
#>  6 1994-11-08 16:00:00 Shahab-1     Kermanshah A… 34.3458         
#>  7 1994-11-08 16:00:00 Shahab-1     Kermanshah A… 34.3458         
#>  8 1997-10-13 17:00:00 Shahab-2     NA            NA              
#>  9 1998-07-21 17:00:00 Shahab-3     NA            NA              
#> 10 1999-06-09 17:00:00 Shahab-1     Kermanshah A… 34.3458         
#> # … with 128 more rows, and 10 more variables: facility_longitude <chr>,
#> #   landing_location <chr>, apogee_km <dbl>, distance_travelled_km <dbl>,
#> #   test_confirmed <lgl>, test_successful <lgl>, additional_information <chr>,
#> #   sources <chr>, missile_family <chr>, in_operational_use <lgl>

dplyr::glimpse(iran_missile_tests_df)

#> Observations: 138
#> Variables: 14
#> $ launch_time            <dttm> 1991-04-30 17:00:00, 1994-11-05 16:00:00, 199…
#> $ missile_name           <chr> "Shahab-2", "Shahab-1", "Shahab-1", "Shahab-1"…
#> $ facility_name          <chr> NA, "Kermanshah Air Base", "Kermanshah Air Bas…
#> $ facility_latitude      <chr> NA, "34.3458", "34.3458", "34.3458", "34.3458"…
#> $ facility_longitude     <chr> NA, "47.1582", "47.1582", "47.1582", "47.1582"…
#> $ landing_location       <chr> NA, "MKO Base \"Asharaf\"", "MKO Base \"Ashara…
#> $ apogee_km              <dbl> 200, NA, NA, NA, NA, NA, NA, NA, 0, NA, NA, NA…
#> $ distance_travelled_km  <dbl> NA, 85, 85, 85, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ test_confirmed         <lgl> TRUE, TRUE, TRUE, TRUE, FALSE, FALSE, FALSE, T…
#> $ test_successful        <lgl> NA, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, …
#> $ additional_information <chr> "First known test of the Shahab 2. The missile…
#> $ sources                <chr> "http://www.astronautix.com/i/iran.html; http:…
#> $ missile_family         <chr> "SRBM", "SRBM", "SRBM", "SRBM", "SRBM", "SRBM"…
#> $ in_operational_use     <lgl> FALSE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, FAL…

# india ===============================================================================
india_missile_tests_df

#> # A tibble: 217 x 11
#>    launch_time         missile_name launch_authority facility_name
#>    <dttm>              <chr>        <chr>            <chr>        
#>  1 1979-08-09 17:00:00 SLV-3        ISRO             Satish Dhawa…
#>  2 1980-07-17 17:00:00 SLV-3        ISRO             Satish Dhawa…
#>  3 1981-05-30 17:00:00 SLV-3        ISRO             Satish Dhawa…
#>  4 1983-04-16 16:00:00 SLV-3        ISRO             Satish Dhawa…
#>  5 1987-03-23 16:00:00 ASLV         ISRO             Satish Dhawa…
#>  6 1988-02-24 16:00:00 Prithvi      IDRDL            Satish Dhawa…
#>  7 1988-07-12 17:00:00 ASLV         ISRO             Satish Dhawa…
#>  8 1989-05-21 17:00:00 Agni-TD      IDRDL            Integrated T…
#>  9 1989-09-26 17:00:00 Prithvi      IDRDL            Satish Dhawa…
#> 10 1991-02-10 16:00:00 Prithvi      IDRDL            Satish Dhawa…
#> # … with 207 more rows, and 7 more variables: landing_location <chr>,
#> #   apogee_km <dbl>, distance_travelled_km <dbl>, test_confirmed <lgl>,
#> #   test_successful <lgl>, additional_information <chr>, sources <chr>

dplyr::glimpse(india_missile_tests_df)

#> Observations: 217
#> Variables: 11
#> $ launch_time            <dttm> 1979-08-09 17:00:00, 1980-07-17 17:00:00, 198…
#> $ missile_name           <chr> "SLV-3", "SLV-3", "SLV-3", "SLV-3", "ASLV", "P…
#> $ launch_authority       <chr> "ISRO", "ISRO", "ISRO", "ISRO", "ISRO", "IDRDL…
#> $ facility_name          <chr> "Satish Dhawan Space Centre", "Satish Dhawan S…
#> $ landing_location       <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ apogee_km              <dbl> 23, 921, 374, 485, 10, 100, 25, 300, 100, 100,…
#> $ distance_travelled_km  <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ test_confirmed         <lgl> TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE…
#> $ test_successful        <lgl> FALSE, TRUE, FALSE, TRUE, FALSE, TRUE, FALSE, …
#> $ additional_information <chr> "India's first space launch", "India's first s…
#> $ sources                <chr> "https://www.indiatoday.in/magazine/science-an…

# pakistan ============================================================================
pakistan_missile_tests_df

#> # A tibble: 95 x 11
#>    launch_time         missile_name launch_authority facility_name
#>    <dttm>              <chr>        <chr>            <chr>        
#>  1 1989-01-31 16:00:00 Hatf-1       SUPA             Sonmiani Ter…
#>  2 1989-01-31 16:00:00 Hatf-2       SUPARCO          Sonmiani Ter…
#>  3 1989-03-31 16:00:00 Hatf-1       NA               Unknown (3)  
#>  4 1991-12-31 16:00:00 Hatf-1       NA               Unknown (3)  
#>  5 1997-06-30 17:00:00 Ghaznavi     SUPARCO          Unknown (3)  
#>  6 1998-04-05 17:00:00 Ghauri-1     KRL              Tilla Satell…
#>  7 1999-01-14 16:00:00 Shaheen-1    PAEC             Sonmiani Ter…
#>  8 1999-04-13 17:00:00 Ghauri-2     KRL              Tilla Satell…
#>  9 1999-04-14 17:00:00 Shaheen-1    PAEC             Sonmiani Ter…
#> 10 2000-01-31 16:00:00 Hatf-1       PK               Sonmiani Ter…
#> # … with 85 more rows, and 7 more variables: landing_location <chr>,
#> #   apogee_km <dbl>, distance_travelled_km <dbl>, test_confirmed <lgl>,
#> #   test_successful <lgl>, additional_information <chr>, sources <chr>

dplyr::glimpse(pakistan_missile_tests_df)

#> Observations: 95
#> Variables: 11
#> $ launch_time            <dttm> 1989-01-31 16:00:00, 1989-01-31 16:00:00, 198…
#> $ missile_name           <chr> "Hatf-1", "Hatf-2", "Hatf-1", "Hatf-1", "Ghazn…
#> $ launch_authority       <chr> "SUPA", "SUPARCO", NA, NA, "SUPARCO", "KRL", "…
#> $ facility_name          <chr> "Sonmiani Terminal", "Sonmiani Terminal", "Unk…
#> $ landing_location       <chr> NA, NA, NA, NA, NA, NA, "Sonmiani Space Facili…
#> $ apogee_km              <dbl> 70, 100, NA, NA, NA, 150, 0, 150, 100, 70, 0, …
#> $ distance_travelled_km  <dbl> NA, NA, NA, NA, NA, NA, 0, 1100, NA, NA, NA, N…
#> $ test_confirmed         <lgl> TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE…
#> $ test_successful        <lgl> TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, FALSE, TRU…
#> $ additional_information <chr> "While the distance of the test was not record…
#> $ sources                <chr> "http://www.astronautix.com/h/hatfi.html https…