The CNS North Korea Missile Test Database

The James Martin Center for Nonproliferation Studies (CNS) North Korea Missile Test Database is the first database to record flight tests of all missiles launched by North Korea capable of delivering a payload of at least 500 kilograms (1102.31 pounds) a distance of at least 300 kilometers (186.4 miles). The database captures advancements in North Korea's missile program by documenting all such tests since the first one occurred in April 1984, and will be routinely updated as events warrant.

dprk_missile_tests_df

Format

An object of class tbl_df (inherits from tbl, data.frame) with 140 rows and 18 columns.

Source

https://www.nti.org/analysis/articles/cns-north-korea-missile-test-database/

References

Created by The James Martin Center for Nonproliferation Studies for the Nuclear Threat Initiative

Examples

dprk_missile_tests_df

#> # A tibble: 140 x 18
#>    date       date_entered_up… launch_time_utc     missile_name missile_type
#>    <date>     <date>           <dttm>              <chr>        <chr>       
#>  1 1984-04-09 2016-12-23       NA                  Scud-B       SRBM        
#>  2 1984-04-09 2016-12-23       NA                  Scud-B       SRBM        
#>  3 1984-04-09 2016-12-23       NA                  Scud-B       SRBM        
#>  4 1984-09-01 2016-12-23       NA                  Scud-B       SRBM        
#>  5 1984-09-01 2016-12-23       NA                  Scud-B       SRBM        
#>  6 1984-09-01 2016-12-23       NA                  Scud-B       SRBM        
#>  7 1986-05-01 2016-12-23       NA                  NA           NA          
#>  8 1990-05-01 2016-12-23       NA                  Nodong       MRBM        
#>  9 1990-06-01 2016-12-23       NA                  Scud-C       SRBM        
#> 10 1991-07-01 2016-12-23       NA                  Scud-C       SRBM        
#> # … with 130 more rows, and 13 more variables: launch_agency_authority <chr>,
#> #   facility_name <chr>, facility_location <chr>, other_name <chr>,
#> #   facility_latitude <dbl>, facility_longitude <dbl>, landing_location <chr>,
#> #   apogee_km <dbl>, distance_travelled_km <dbl>, confirmed <lgl>,
#> #   success <lgl>, additional_information <chr>, sources <chr>

dplyr::glimpse(dprk_missile_tests_df)

#> Observations: 140
#> Variables: 18
#> $ date                    <date> 1984-04-09, 1984-04-09, 1984-04-09, 1984-09-…
#> $ date_entered_updated    <date> 2016-12-23, 2016-12-23, 2016-12-23, 2016-12-…
#> $ launch_time_utc         <dttm> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ missile_name            <chr> "Scud-B", "Scud-B", "Scud-B", "Scud-B", "Scud…
#> $ missile_type            <chr> "SRBM", "SRBM", "SRBM", "SRBM", "SRBM", "SRBM…
#> $ launch_agency_authority <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ facility_name           <chr> "Tonghae Satellite Launching Ground", "Tongha…
#> $ facility_location       <chr> "Hwadae County, North Hamgyong Province", "Hw…
#> $ other_name              <chr> "Musudan-ri", "Musudan-ri", "Musudan-ri", "Mu…
#> $ facility_latitude       <dbl> 40.85000, 40.85000, 40.85000, 40.85000, 40.85…
#> $ facility_longitude      <dbl> 129.6667, 129.6667, 129.6667, 129.6667, 129.6…
#> $ landing_location        <chr> NA, NA, NA, NA, NA, NA, NA, NA, "Sea of Japan…
#> $ apogee_km               <dbl> NA, NA, 200, NA, 200, 200, NA, 150, 200, NA, …
#> $ distance_travelled_km   <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 5…
#> $ confirmed               <lgl> TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, FALSE, FA…
#> $ success                 <lgl> TRUE, TRUE, TRUE, FALSE, FALSE, FALSE, NA, FA…
#> $ additional_information  <chr> "First known test by North Korea of a missile…
#> $ sources                 <chr> "http://www.nti.org/media/pdfs/north_korea_mi…