Nuclear Cooperation Agreement (NCA) Dataset

The NCA dataset contains information on nuclear cooperation agreements signed by countries from 1945 to 2003. It includes more than 1,500 treaties in which states pledged to transfer nuclear technology, materials, or know-how. The dataset is based on the 2003 version of James Keeley's list of nuclear cooperation agreements.

nca_df

Format

An object of class tbl_df (inherits from tbl, data.frame) with 3152 rows and 7 columns.

Source

http://www.matthewfuhrmann.com/datasets.html

References

Fuhrmann, Matthew. 2012. Atomic Assistance: How "Atoms for Peace" Programs Cause Nuclear Insecurity. Ithaca, NY: Cornell University Press.

Examples

nca_df

#> # A tibble: 3,152 x 7
#>    ccode1 ccode2 country1 country2       year   nca nca_type                    
#>     <dbl>  <dbl> <chr>    <chr>         <dbl> <dbl> <chr>                       
#>  1      2     20 USA      Canada         1955     1 comprehensive power, restri…
#>  2      2     20 USA      Canada         1969     1 intangibles                 
#>  3      2     20 USA      Canada         1987     1 intangibles                 
#>  4      2     20 USA      Canada         1987     1 intangibles                 
#>  5      2     20 USA      Canada         1988     1 intangibles                 
#>  6      2     40 USA      Cuba           1956     1 research                    
#>  7      2     42 USA      Dominican Rep  1956     1 research                    
#>  8      2     70 USA      Mexico         1965     1 nuclear safety              
#>  9      2     70 USA      Mexico         1969     1 intangibles                 
#> 10      2     70 USA      Mexico         1980     1 nuclear safety              
#> # … with 3,142 more rows

dplyr::glimpse(nca_df)

#> Observations: 3,152
#> Variables: 7
#> $ ccode1   <dbl> 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, …
#> $ ccode2   <dbl> 20, 20, 20, 20, 20, 40, 42, 70, 70, 70, 70, 70, 70, 70, 90, …
#> $ country1 <chr> "USA", "USA", "USA", "USA", "USA", "USA", "USA", "USA", "USA…
#> $ country2 <chr> "Canada", "Canada", "Canada", "Canada", "Canada", "Cuba", "D…
#> $ year     <dbl> 1955, 1969, 1987, 1987, 1988, 1956, 1956, 1965, 1969, 1980, …
#> $ nca      <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, …
#> $ nca_type <chr> "comprehensive power, restricted", "intangibles", "intangibl…