Nuclear Latency (NL) Dataset

The NL dataset provides information on the global diffusion of latent nuclear capabilities from 1939 to 2012. It identifies all known uranium enrichment and plutonium reprocessing facilities that have been built globally during the nuclear age.

nl_1.2_df

Format

An object of class tbl_df (inherits from tbl, data.frame) with 253 rows and 20 columns.

Source

http://www.matthewfuhrmann.com/datasets.html

References

Fuhrmann, Matthew and Benjamin Tkach. 2015. "Almost Nuclear: Introducing the Nuclear Latency Dataset." Conflict Management and Peace Science 32 (4): 443-461.

Examples

nl_1.2_df

#> # A tibble: 253 x 20
#>    country_name ccode facility_name facility_ambigu… enr_type size 
#>    <chr>        <dbl> <chr>         <lgl>            <chr>    <chr>
#>  1 Algeria        615 Hot cell fac… FALSE            reproce… labo…
#>  2 Argentina      160 Ezeiza – SF … FALSE            reproce… labo…
#>  3 Argentina      160 Ezeiza II – … NA               reproce… pilot
#>  4 Argentina      160 Pilcaniyeu E… FALSE            gaseous… pilot
#>  5 Argentina      160 Pilcaniyeu E… NA               gaseous… pilot
#>  6 Australia      900 Lucas Heights FALSE            centrif… labo…
#>  7 Australia      900 Silex Laser … FALSE            laser    labo…
#>  8 Belgium        211 Mol, Purex R… FALSE            reproce… pilot
#>  9 Brazil         140 Aerospace Te… FALSE            laser    labo…
#> 10 Brazil         140 BRF Enrichme… FALSE            centrif… pilot
#> # … with 243 more rows, and 14 more variables: construction_start <dbl>,
#> #   construction_end <dbl>, operation_start <dbl>, operation_end <dbl>,
#> #   operation2_start <dbl>, operation2_end <dbl>, covert <lgl>, iaea <lgl>,
#> #   regional <lgl>, military <lgl>, military_ambiguity <lgl>,
#> #   multinational <lgl>, foreign_assistance <lgl>,
#> #   foreign_assistance_ambiguity <lgl>

dplyr::glimpse(nl_1.2_df)

#> Observations: 253
#> Variables: 20
#> $ country_name                 <chr> "Algeria", "Argentina", "Argentina", "Ar…
#> $ ccode                        <dbl> 615, 160, 160, 160, 160, 900, 900, 211, …
#> $ facility_name                <chr> "Hot cell facility at Ain Oussera", "Eze…
#> $ facility_ambiguity           <lgl> FALSE, FALSE, NA, FALSE, NA, FALSE, FALS…
#> $ enr_type                     <chr> "reprocessing", "reprocessing", "reproce…
#> $ size                         <chr> "laboratory", "laboratory", "pilot", "pi…
#> $ construction_start           <dbl> 1986, 1968, 1978, 1979, 2000, 1965, 1982…
#> $ construction_end             <dbl> 1992, 1968, 1990, 1987, NA, 1972, 1992, …
#> $ operation_start              <dbl> 1992, 1968, NA, 1987, NA, 1972, 1992, 19…
#> $ operation_end                <dbl> NA, 1973, NA, 1994, NA, 1983, 2007, 1974…
#> $ operation2_start             <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ operation2_end               <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ covert                       <lgl> TRUE, FALSE, FALSE, TRUE, FALSE, FALSE, …
#> $ iaea                         <lgl> TRUE, FALSE, FALSE, FALSE, TRUE, TRUE, T…
#> $ regional                     <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE…
#> $ military                     <lgl> FALSE, FALSE, FALSE, TRUE, FALSE, FALSE,…
#> $ military_ambiguity           <lgl> TRUE, TRUE, FALSE, FALSE, FALSE, TRUE, F…
#> $ multinational                <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE…
#> $ foreign_assistance           <lgl> TRUE, FALSE, FALSE, FALSE, FALSE, FALSE,…
#> $ foreign_assistance_ambiguity <lgl> FALSE, TRUE, TRUE, TRUE, TRUE, FALSE, TR…