UCDP Georeferenced Event Dataset (GED) Global version 19.1

This dataset is UCDP's most disaggregated dataset, covering individual events of organized violence (phenomena of lethal violence occurring at a given time and place). These events are sufficiently fine-grained to be geo-coded down to the level of individual villages, with temporal durations disaggregated to single, individual days.

ucdp_19.1_df

Format

An object of class tbl_df (inherits from tbl, data.frame) with 152616 rows and 42 columns.

Source

https://ucdp.uu.se/downloads/index.html#ged_global/

References

Sundberg, Ralph, and Erik Melander, 2013, “Introducing the UCDP Georeferenced Event Dataset”, Journal of Peace Research, vol.50, no.4, 523-532

Högbladh Stina, 2019, “UCDP GED Codebook version 19.1”, Department of Peace and Conflict Research, Uppsala University

Examples

ucdp_19.1_df

#> # A tibble: 152,616 x 42
#>        id  year active_year type_of_violence conflict_new_id conflict_name
#>     <dbl> <dbl>       <dbl>            <dbl>           <dbl> <chr>        
#>  1  67972  2013           1                1           11347 Mali: Govern…
#>  2  23385  2004           1                3             583 FNI - Civili…
#>  3  24255  2007           0                2            4600 CNDP - PARECO
#>  4  82612  2008           1                1             364 India: Kashm…
#>  5  82645  2008           1                1             364 India: Kashm…
#>  6 158618  2011           1                1             354 Turkey: Kurd…
#>  7  87396  1997           0                3             523 ULFA - Civil…
#>  8  60774  2013           1                1             283 DR Congo (Za…
#>  9  25360  2010           1                1             337 Somalia: Gov…
#> 10  11995  2003           1                1             288 Chad: Govern…
#> # … with 152,606 more rows, and 36 more variables: dyad_new_id <dbl>,
#> #   dyad_name <chr>, side_a_new_id <dbl>, gwnoa <dbl>, side_a <chr>,
#> #   side_b_new_id <dbl>, gwnob <dbl>, side_b <chr>, number_of_sources <dbl>,
#> #   source_article <chr>, source_office <chr>, source_date <chr>,
#> #   source_headline <chr>, source_original <chr>, where_prec <dbl>,
#> #   where_coordinates <chr>, adm_1 <chr>, adm_2 <chr>, latitude <dbl>,
#> #   longitude <dbl>, geom_wkt <chr>, priogrid_gid <dbl>, country <chr>,
#> #   country_id <dbl>, region <chr>, event_clarity <dbl>, date_prec <dbl>,
#> #   date_start <date>, date_end <date>, deaths_a <dbl>, deaths_b <dbl>,
#> #   deaths_civilians <dbl>, deaths_unknown <dbl>, low <dbl>, best <dbl>,
#> #   high <dbl>

dplyr::glimpse(ucdp_19.1_df)

#> Observations: 152,616
#> Variables: 42
#> $ id                <dbl> 67972, 23385, 24255, 82612, 82645, 158618, 87396, 6…
#> $ year              <dbl> 2013, 2004, 2007, 2008, 2008, 2011, 1997, 2013, 201…
#> $ active_year       <dbl> 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, …
#> $ type_of_violence  <dbl> 1, 3, 2, 1, 1, 1, 3, 1, 1, 1, 1, 3, 1, 1, 1, 1, 3, …
#> $ conflict_new_id   <dbl> 11347, 583, 4600, 364, 364, 354, 523, 283, 337, 288…
#> $ conflict_name     <chr> "Mali: Government", "FNI - Civilians", "CNDP - PARE…
#> $ dyad_new_id       <dbl> 12571, 1050, 5210, 792, 792, 781, 990, 10509, 750, …
#> $ dyad_name         <chr> "Government of Mali - MUJAO", "FNI - Civilians", "C…
#> $ side_a_new_id     <dbl> 72, 606, 426, 141, 141, 115, 326, 89, 95, 87, 112, …
#> $ gwnoa             <dbl> 432, NA, NA, 750, 750, 640, NA, 490, 520, 483, 625,…
#> $ side_a            <chr> "Government of Mali", "FNI", "CNDP", "Government of…
#> $ side_b_new_id     <dbl> 1161, 1, 896, 325, 325, 323, 1, 1200, 717, 452, 469…
#> $ gwnob             <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ side_b            <chr> "MUJAO", "Civilians", "PARECO", "Kashmir insurgents…
#> $ number_of_sources <dbl> 7, -1, -1, -1, -1, -1, -1, 8, -1, -1, -1, -1, -1, -…
#> $ source_article    <chr> "\"Agence France Presse,2013-01-12,French pilot kil…
#> $ source_office     <chr> "Agence France Presse;Agence France Presse;Agence F…
#> $ source_date       <chr> "2013-01-12;2013-01-28;2013-01-12;2013-01-12;2013-0…
#> $ source_headline   <chr> "French pilot killed in Mali helicopter raid: defen…
#> $ source_original   <chr> "French Defence Minister Jean-Yves Le Drian;  resid…
#> $ where_prec        <dbl> 1, 1, 2, 1, 3, 3, 4, 1, 1, 1, 4, 2, 3, 2, 3, 1, 1, …
#> $ where_coordinates <chr> "Konna town", "Gobu village", "Ngungu village", "Ch…
#> $ adm_1             <chr> "Mopti region", "Ituri province", "Nord Kivu provin…
#> $ adm_2             <chr> "Mopti cercle", "Djugu territory", "Masisi territor…
#> $ latitude          <dbl> 14.943290, 1.769722, -1.651944, 34.378934, 34.37260…
#> $ longitude         <dbl> -3.89474, 30.77583, 28.87694, 74.72549, 74.16729, 4…
#> $ geom_wkt          <chr> "POINT (-3.894740 14.943290)", "POINT (30.775833 1.…
#> $ priogrid_gid      <dbl> 150833, 132182, 127138, 179070, 179069, 184761, 167…
#> $ country           <chr> "Mali", "DR Congo (Zaire)", "DR Congo (Zaire)", "In…
#> $ country_id        <dbl> 432, 490, 490, 750, 750, 640, 750, 490, 520, 483, 6…
#> $ region            <chr> "Africa", "Africa", "Africa", "Asia", "Asia", "Midd…
#> $ event_clarity     <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, …
#> $ date_prec         <dbl> 2, 2, 2, 2, 2, 2, 1, 2, 2, 2, 2, 2, 2, 2, 2, 2, 1, …
#> $ date_start        <date> 2013-01-11, 2004-01-14, 2007-12-13, 2008-03-04, 20…
#> $ date_end          <date> 2013-01-12, 2004-01-16, 2007-12-18, 2008-03-05, 20…
#> $ deaths_a          <dbl> 12, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ deaths_b          <dbl> 16, 0, 0, 2, 11, 13, 0, 0, 0, 2, 2, 0, 8, 0, 10, 0,…
#> $ deaths_civilians  <dbl> 3, 200, 4, 0, 0, 0, 1, 0, 0, 0, 0, 6, 0, 0, 0, 0, 1…
#> $ deaths_unknown    <dbl> 0, 0, 0, 0, 0, 0, 0, 80, 11, 0, 0, 0, 17, 0, 0, 16,…
#> $ low               <dbl> 31, 100, 4, 2, 12, 13, 1, 70, 11, 2, 2, 6, 25, 0, 1…
#> $ best              <dbl> 31, 200, 4, 2, 12, 13, 1, 80, 11, 2, 2, 6, 25, 0, 1…
#> $ high              <dbl> 50, 200, 4, 2, 12, 30, 1, 80, 11, 72, 19, 6, 25, 7,…