UCDP Georeferenced Event Dataset (GED) Syria 20160101-20191130 (652.1601.1911)

This dataset contains georeferenced events data for Syria, covering 20160101-20191130. The dataset includes the same variables as the UCDP GED Global 19.1 dataset, and the same Codebook applies to both datasets. In contrast to UCDP GED Global 19.1, this dataset includes all clear events, meaning that dyads do not have to pass the threshold of 25 battle-related deaths during one calendar year, to be included (see the UCDP GED Codebook for definitions and more information on events, dyads, and the 25 threshold).

ucdp_syria_df

Format

An object of class tbl_df (inherits from tbl, data.frame) with 16566 rows and 41 columns.

Source

https://ucdp.uu.se/downloads/index.html#ged_global/

References

Sundberg, Ralph, and Erik Melander, 2013, “Introducing the UCDP Georeferenced Event Dataset”, Journal of Peace Research, vol.50, no.4, 523-532

Högbladh Stina, 2019, “UCDP GED Codebook version 19.1”, Department of Peace and Conflict Research, Uppsala University

Examples

ucdp_syria_df

#> # A tibble: 16,566 x 41
#>        id  year active_year type_of_violence conflict_new_id conflict_name
#>     <dbl> <dbl>       <dbl>            <dbl>           <dbl> <chr>        
#>  1 214797  2016           1                1             299 Syria: Gover…
#>  2 214804  2016           1                1             299 Syria: Gover…
#>  3 214808  2016           1                1             299 Syria: Gover…
#>  4 214811  2016           1                1             299 Syria: Gover…
#>  5 214814  2016           1                2           13935 Ahrar al-Sha…
#>  6 214817  2016           1                2           13935 Ahrar al-Sha…
#>  7 214825  2016           1                1             299 Syria: Gover…
#>  8 214828  2016           1                1             299 Syria: Gover…
#>  9 214831  2016           1                1             299 Syria: Gover…
#> 10 214858  2016           1                1             299 Syria: Gover…
#> # … with 16,556 more rows, and 35 more variables: dyad_new_id <dbl>,
#> #   dyad_name <chr>, side_a_new_id <dbl>, side_a <chr>, side_b_new_id <dbl>,
#> #   gwnob <dbl>, side_b <chr>, number_of_sources <dbl>, source_article <chr>,
#> #   source_office <chr>, source_date <chr>, source_headline <chr>,
#> #   source_original <chr>, where_prec <dbl>, where_coordinates <chr>,
#> #   adm_1 <chr>, adm_2 <chr>, latitude <dbl>, longitude <dbl>, geom_wkt <chr>,
#> #   priogrid_gid <dbl>, country <chr>, country_id <dbl>, region <chr>,
#> #   event_clarity <dbl>, date_prec <dbl>, date_start <date>, date_end <date>,
#> #   deaths_a <dbl>, deaths_b <dbl>, deaths_civilians <dbl>,
#> #   deaths_unknown <dbl>, low <dbl>, best <dbl>, high <dbl>

dplyr::glimpse(ucdp_syria_df)

#> Observations: 16,566
#> Variables: 41
#> $ id                <dbl> 214797, 214804, 214808, 214811, 214814, 214817, 214…
#> $ year              <dbl> 2016, 2016, 2016, 2016, 2016, 2016, 2016, 2016, 201…
#> $ active_year       <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, …
#> $ type_of_violence  <dbl> 1, 1, 1, 1, 2, 2, 1, 1, 1, 1, 2, 1, 1, 1, 1, 1, 1, …
#> $ conflict_new_id   <dbl> 299, 299, 299, 299, 13935, 13935, 299, 299, 299, 29…
#> $ conflict_name     <chr> "Syria: Government", "Syria: Government", "Syria: G…
#> $ dyad_new_id       <dbl> 11973, 11973, 11973, 11973, 15089, 15089, 11973, 11…
#> $ dyad_name         <chr> "Government of Syria - Syrian insurgents", "Governm…
#> $ side_a_new_id     <dbl> 118, 118, 118, 118, 6389, 6389, 118, 118, 118, 118,…
#> $ side_a            <chr> "Government of Syria", "Government of Syria", "Gove…
#> $ side_b_new_id     <dbl> 4456, 4456, 4456, 4456, 6288, 6288, 4456, 4456, 445…
#> $ gwnob             <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ side_b            <chr> "Syrian insurgents", "Syrian insurgents", "Syrian i…
#> $ number_of_sources <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 2, 3, …
#> $ source_article    <chr> "\"SOHR,2015-01-01,14 shells hit Dahya al-Assad an …
#> $ source_office     <chr> "SOHR", "SOHR", "SOHR", "SOHR", "SOHR", "SOHR", "SO…
#> $ source_date       <chr> "2015-01-01", "2016-01-01", "2016-01-01", "2016-01-…
#> $ source_headline   <chr> "14 shells hit Dahya al-Assad an clashes continue a…
#> $ source_original   <chr> "SOHR", "SOHR", "SOHR", "SOHR", "SOHR", "SOHR", "SO…
#> $ where_prec        <dbl> 1, 2, 1, 2, 2, 2, 2, 2, 3, 2, 2, 1, 2, 1, 1, 2, 1, …
#> $ where_coordinates <chr> "Teir Maaleh village", "Ein Laroz town", "Aleppo to…
#> $ adm_1             <chr> "Homs Governorate", "Idlib Governorate", "Aleppo Go…
#> $ adm_2             <chr> "Ar Rastan district", "Ariha district", "Jabal Sama…
#> $ latitude          <dbl> 34.80246, 35.68834, 36.20124, 32.82944, 36.49613, 3…
#> $ longitude         <dbl> 36.70760, 36.44996, 37.16117, 36.15933, 37.00332, 3…
#> $ geom_wkt          <chr> "POINT (36.707602 34.802462)", "POINT (36.449962 35…
#> $ priogrid_gid      <dbl> 179714, 181153, 181875, 176833, 181875, 181875, 181…
#> $ country           <chr> "Syria", "Syria", "Syria", "Syria", "Syria", "Syria…
#> $ country_id        <dbl> 652, 652, 652, 652, 652, 652, 652, 652, 652, 652, 6…
#> $ region            <chr> "Middle East", "Middle East", "Middle East", "Middl…
#> $ event_clarity     <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, …
#> $ date_prec         <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, …
#> $ date_start        <date> 2016-01-01, 2016-01-01, 2016-01-01, 2016-01-01, 20…
#> $ date_end          <date> 2016-01-01, 2016-01-01, 2016-01-01, 2016-01-01, 20…
#> $ deaths_a          <dbl> 2, 0, 2, 6, 1, 0, 2, 1, 2, 0, 2, 2, 0, 0, 0, 0, 0, …
#> $ deaths_b          <dbl> 2, 0, 0, 1, 1, 2, 0, 0, 0, 0, 2, 0, 0, 2, 0, 22, 0,…
#> $ deaths_civilians  <dbl> 0, 2, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 6, 0, 1, 6, 5, …
#> $ deaths_unknown    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ low               <dbl> 4, 2, 2, 7, 2, 2, 2, 1, 2, 1, 4, 2, 6, 2, 1, 28, 5,…
#> $ best              <dbl> 4, 2, 2, 7, 2, 2, 2, 1, 2, 1, 4, 2, 6, 2, 1, 28, 5,…
#> $ high              <dbl> 4, 2, 2, 7, 2, 2, 2, 1, 2, 1, 4, 2, 6, 2, 1, 28, 5,…