Spaniel and Smith's 𝜈-CLEAR Dataset

Abstract: The causes and consequences of nuclear proficiency are central to important questions in international relations. At present, researchers tend to use observable characteristics as a proxy. However, aggregation is a problem: existing measures implicitly assume that each indicator is equally informative and that measurement error is not a concern. We overcome these issues by applying a statistical measurement model to directly estimate nuclear proficiency from observed indicators. The resulting estimates form a new dataset on nuclear proficiency which we call 𝜈-CLEAR. We demonstrate that these estimates are consistent with known patterns of nuclear proficiency while also uncovering more nuance than existing measures. Additionally, we demonstrate how scholars can use these estimates to account for measurement error by revisiting existing results with our measure.

v_clear_df

Format

An object of class tbl_df (inherits from tbl, data.frame) with 8086 rows and 20 columns.

Source

https://williamspaniel.com/papers/nuclear/

References

Smith, Bradley & Spaniel, William. (2018). Introducing ν-CLEAR: a latent variable approach to measuring nuclear proficiency. Conflict Management and Peace Science. 073889421774161. 10.1177/0738894217741619.

Examples

v_clear_df

#> # A tibble: 8,086 x 20
#>       id ccode country_name  year uranium_possess… metallurgical_c…
#>    <int> <int> <chr>        <int> <lgl>            <lgl>           
#>  1 21938     2 United Stat…  1938 TRUE             TRUE            
#>  2 21939     2 United Stat…  1939 TRUE             TRUE            
#>  3 21940     2 United Stat…  1940 TRUE             TRUE            
#>  4 21941     2 United Stat…  1941 TRUE             TRUE            
#>  5 21942     2 United Stat…  1942 TRUE             TRUE            
#>  6 21943     2 United Stat…  1943 TRUE             TRUE            
#>  7 21944     2 United Stat…  1944 TRUE             TRUE            
#>  8 21945     2 United Stat…  1945 TRUE             TRUE            
#>  9 21946     2 United Stat…  1946 TRUE             TRUE            
#> 10 21947     2 United Stat…  1947 TRUE             TRUE            
#> # … with 8,076 more rows, and 14 more variables: chemical_capability <lgl>,
#> #   nitric_production <lgl>, electricity_production <lgl>,
#> #   nuclear_engineering <lgl>, explosive_production <lgl>,
#> #   non_heavy_water_reactor <lgl>, heavy_water_reactor <lgl>,
#> #   nuclear_test <lgl>, reprocessing <lgl>, uranium_enrichment <lgl>,
#> #   submarines <lgl>, weapons_exploration <lgl>, weapons_pursuit <lgl>,
#> #   nuclear_weapons <lgl>

dplyr::glimpse(v_clear_df)

#> Observations: 8,086
#> Variables: 20
#> $ id                       <int> 21938, 21939, 21940, 21941, 21942, 21943, 21…
#> $ ccode                    <int> 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2,…
#> $ country_name             <chr> "United States of America", "United States o…
#> $ year                     <int> 1938, 1939, 1940, 1941, 1942, 1943, 1944, 19…
#> $ uranium_possession       <lgl> TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TR…
#> $ metallurgical_capability <lgl> TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TR…
#> $ chemical_capability      <lgl> TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TR…
#> $ nitric_production        <lgl> TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TR…
#> $ electricity_production   <lgl> TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TR…
#> $ nuclear_engineering      <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FA…
#> $ explosive_production     <lgl> TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TR…
#> $ non_heavy_water_reactor  <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FA…
#> $ heavy_water_reactor      <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FA…
#> $ nuclear_test             <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FA…
#> $ reprocessing             <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, TRUE, TRU…
#> $ uranium_enrichment       <lgl> FALSE, FALSE, FALSE, TRUE, TRUE, TRUE, TRUE,…
#> $ submarines               <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FA…
#> $ weapons_exploration      <lgl> FALSE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, T…
#> $ weapons_pursuit          <lgl> FALSE, FALSE, FALSE, FALSE, TRUE, TRUE, TRUE…
#> $ nuclear_weapons          <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FA…